requirements.txt
Django==4.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.6
redis==4.5.5
# Optional for tasks:
celery==5.2.7
# Optional for DRF:
djangorestframework==3.14.0
# For 2FA:
django-two-factor-auth==1.14.0
django-otp==1.2.0
phonenumbers==8.13.13
requests==2.31.0
django-adminlte3
psutil==5.9.7
pillow
drf-yasg>=1.20.0
